First pass on reorganizing functions
